fix warning

+++ zergtool/example/write_fst.cpp
@@ -4,14 +4,14 @@ using namespace ztool;
 
 int main() {
     std::string path_ = "/tmp/data2.fst";
-    int nrOfRows = 10;
+    size_t nrOfRows = 10;
 
     std::vector<double> d_vec(nrOfRows);
     std::vector<float> f_vec(nrOfRows);
     std::vector<int> i_vec(nrOfRows);
     std::vector<std::string> s_vec(nrOfRows);
     std::vector<bool> b_vec(nrOfRows);
-    for (int i = 0; i < nrOfRows; ++i) {
+    for (size_t i = 0; i < nrOfRows; ++i) {
         d_vec[i] = i;
         f_vec[i] = i;
         i_vec[i] = i;
